remove object-based memory cacher, rename memory-map cacher to memory
