+++ src/cachers/memory.js
@@ -11,25 +11,25 @@ const nanomatch  	= require("nanomatch");
 const BaseCacher  	= require("./base");
 /**
  * Cacher factory for memory cache
- *
+ * 
  * 		Similar: https://github.com/mpneuried/nodecache/blob/master/_src/lib/node_cache.coffee
- *
- * @class MemoryCacher
+ * 
+ * @class MemoryMapCacher
  */
-class MemoryCacher extends BaseCacher {
+class MemoryMapCacher extends BaseCacher {
 
 	/**
-	 * Creates an instance of MemoryCacher.
-	 *
+	 * Creates an instance of MemoryMapCacher.
+	 * 
 	 * @param {object} opts
-	 *
-	 * @memberOf MemoryCacher
+	 * 
+	 * @memberOf MemoryMapCacher
 	 */
 	constructor(opts) {
 		super(opts);
 
 		// Cache container
-		this.cache = {};
+		this.cache = new Map();
 
 		if (this.opts.ttl) {
 			this.timer = setInterval(() => {
@@ -43,17 +43,19 @@ class MemoryCacher extends BaseCacher {
 
 	/**
 	 * Get data from cache by key
-	 *
+	 * 
 	 * @param {any} key
 	 * @returns {Promise}
-	 *
-	 * @memberOf Cacher
+	 *  
+	 * @memberOf MemoryMapCacher
 	 */
 	get(key) {
 		//this.logger.debug(`GET ${key}`);
-		let item = this.cache[key];
-		if (item) {
+
+		if (this.cache.has(key)) {
 			//this.logger.debug(`FOUND ${key}`);
+
+			let item = this.cache.get(key);
 
 			if (this.opts.ttl) {
 				// Update expire time (hold in the cache if we are using it)
@@ -66,33 +68,33 @@ class MemoryCacher extends BaseCacher {
 
 	/**
 	 * Save data to cache by key
-	 *
+	 * 
 	 * @param {any} key
 	 * @param {any} data JSON object
 	 * @returns {Promise}
-	 *
-	 * @memberOf Cacher
+	 * 
+	 * @memberOf MemoryMapCacher
 	 */
 	set(key, data) {
-		this.cache[key] = {
+		this.cache.set(key, {
 			data,
 			expire: this.opts.ttl ? Date.now() + this.opts.ttl * 1000 : null
-		};
+		});
 		this.logger.debug(`SET ${key}`);
 		return Promise.resolve(data);
 	}
 
 	/**
 	 * Delete a key from cache
-	 *
+	 * 
 	 * @param {any} key
 	 * @returns {Promise}
-	 *
-	 * @memberOf Cacher
+	 * 
+	 * @memberOf MemoryMapCacher
 	 */
 	del(key) {
-		delete this.cache[key];
-		this.logger.debug(`DELETE ${key}`);
+		this.cache.delete(key);
+		this.logger.debug(`REMOVE ${key}`);
 		return Promise.resolve();
 	}
 
@@ -100,42 +102,38 @@ class MemoryCacher extends BaseCacher {
 	 * Clean cache. Remove every key by match
 	 * @param {any} match string. Default is "**"
 	 * @returns {Promise}
-	 *
+	 * 
 	 * @memberOf Cacher
 	 */
 	clean(match = "**") {
 		this.logger.debug(`CLEAN ${match}`);
 
-		let keys = Object.keys(this.cache);
-		keys.forEach((key) => {
+		this.cache.forEach((value, key) => {
 			if (nanomatch.isMatch(key, match)) {
 				this.logger.debug(`REMOVE ${key}`);
-				delete this.cache[key];
+				this.cache.delete(key);
 			}
 		});
 
 		return Promise.resolve();
 	}
 
-
 	/**
 	 * Check & remove the expired cache items
-	 *
+	 * 
 	 * @memberOf MemoryMapCacher
 	 */
 	checkTTL() {
-		let self = this;
 		let now = Date.now();
-		let keys = Object.keys(this.cache);
-		keys.forEach((key) => {
-			let item = this.cache[key];
+		this.cache.forEach((value, key) => {
+			let item = this.cache.get(key);
 
 			if (item.expire && item.expire < now) {
 				this.logger.debug(`EXPIRED ${key}`);
-				delete self.cache[key];
+				this.cache.delete(key);
 			}
 		});
 	}
 }
 
-module.exports = MemoryCacher;
+module.exports = MemoryMapCacher;
